Consider the night shift ending time for the day coverage of RCs
By taking the end of a night shift into account by checking if the
specific day is covered by run coordinators, cases like the last shift
with no specific run coordinator are unsubscribable are fixed.
A better way would be a time overlap calculation of the particular shift
and corresponding taken run coordinator shifts, which might be done in
the future. So far the condition is looser for fulfilling the RC
requirement for shift subscription.